Lab Task 13/requirements.txt
Flask
python-dotenv
groq
Werkzeug
gunicorn  # optional for production
fpdf2